scripts/requirements.txt
requests
pandas
numpy
psycopg2-binary
sqlalchemy
tenacity
# Apache Airflow - only needed for parsing DAGs
# For Python 3.11: apache-airflow==2.8.0
# For Python 3.12+: apache-airflow>=3.1.6
# Install manually based on your Python version
